# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RIDEXPLORERS API",
    "version": "1.0.0",
    "description": "API providing data about roller coasters and theme parks"
  },
  "servers": [
    { "url": "http://localhost:8000" }
  ],
  "paths": {
...
    "/scrape/start": {
      "get": {
        "summary": "Start scraping script",
        "parameters": [{"name": "script", "in": "query", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Scraping started"}, "400": {"description": "Error"}}
      }
    },
...
    "/scrape/files/{name}": {
      "get": {
        "summary": "Get scraped file",
        "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "File content"}, "404": {"description": "Not found"}}
      }
    },
    "/scrape/upload": {
      "post": {
        "summary": "Upload JSON file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {"type": "string", "format": "binary"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "File uploaded"}, "400": {"description": "Error"}}
      }
    },
...
    "/scrape/logs": {
      "get": {
        "summary": "Retrieve scraping logs",
        "parameters": [{"name": "id", "in": "query", "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Logs"}}
      }
    }
  }
}